requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
langchain==0.1.9
langchain-openai==0.0.8
langchain-core==0.1.26
langgraph==0.0.24
qdrant-client==1.7.3
redis==5.0.1
psycopg2-binary==2.9.9
httpx==0.26.0
python-dotenv==1.0.1
requests==2.31.0
sqlalchemy==2.0.27
greenlet==3.0.3
jinja2==3.1.3
langchain-google-genai==0.0.9
